error handling: introduce ReadySetError & helpers
The new `ReadySetError` type encapsulates pretty much *all* failure
conditions that occur in the main `noria*` crates, and is used
extensively by the other commits in the error handling patch series.
It makes use of the `thiserror` crate to derive `std::error::Error` for
itself.

A number of helper functions and macros are supplied that generate
errors of this type. When refactoring ReadySet code to use proper error
handling, the following substitutions should be made:

- `assert!` should be replaced with `invariant!`
- `assert_eq!` should be replaced with `invariant_eq!`
- `panic!` should be replaced with `internal!` or `unsupported!`:
  - `internal!` is for code paths we hope are never hit
  - `unsupported!` is for code paths that are trigger-able by user
    action, and signals that the requested operation is not supported
    yet.
- `.unwrap()` and `.expect()` should be replaced with
  `.ok_or_else(|| internal_err(/* description */))?`
  - Note that this currently does NOT result in file and line info
    being included in the generated error.

This of course entails converting a *lot* of code into code that
returns `ReadySetResult<T>` where it used to return T, which is done
in other commits in this patch series (!).