error handling: make Controller return ReadySetResult
The Controller type now returns ReadySetResult instead of anyhow::Error,
and the ControllerHandle type has been updated to match.

It didn't make sense for ControllerHandle::rpc to have a third `err`
parameter given the existence of the rpc_err!() macro, so it was taken
out (most of the uses were basically saying "xyz RPC call failed", which
is done for you by rpc_err!()).

`noria-client` is again tweaked to support the change by adding various
.into() calls about the place.

META: At this point in the patch series, the `noria` crate is fully
refactored. Yay!