error handling: refactor the materialization planner
The materialization planner code
(`src/controller/migrate/materialization`), as well as the
`provenance_of()` function, is now error-handling aware. (This undoes a
lot of the unwrapping in that code added by previous changes in this
patchset.)

The migration controller just unwraps the result returned by the
planner, but that will soon change.