Refactor mysql backend to not use ParamParser
Since the backend will now support postgres, we need the implementation
to be db agnostic. Therefore, it is more approporiate for the
implementation to operate on Vec<DataType> instead of a param parser,
which is based on the mysql binary protocol.

Note, the only less than ideal change here is that now
mysql_connector::on_execute() now takes in a Vec<DataType>. Currently
the function is not fully implemented but now, once it is, we will need
to convert the values into mysql values rather than datatype. However I
think the longer term solution is going to be a more robust module for
each supported db within noria-client that handles all of the
conversions.