error handling: refactor the rest of the controller
The `ControllerInner`, and `controller::main`, functions, are now
error-handling aware.

As part of this work, some ...interesting decisions were made on how to
handle recipes failing to apply. Before, recipes failing took down the
whole controller; now, that doesn't happen, but some interesting hacks
are in place to just about keep it together that I'm not sure we want to
keep for very long. The changes made do, however, let you issue a bad
query and have things not come crashing down, so that's something.

META: This completes the refactor of `noria-server`. Yay!