error handling: refactor `noria-client`
`noria-client`'s error type was completely overhauled to use `thiserror`
and use the `ReadySetError` for most of what it does. The crate also
makes use of the error-handling macros from the `noria` crate to keep
things consistent.

In addition, some asserts were removed in various pieces of
backend/connector code.

This tops off the error handling changes, allowing consumers to actually
properly consume the rich error types returned by the rest of the
refactored codebase.

One FIXME left in this commit is the mapping of errors to MySQL error
codes, which is still to be done.