Move query analysis utils into nom-sql
The functions that we had in noria_server::controller::sql::query_utils
are generally useful to anything that consumes queries - specifically,
mir would like to be able to ask which columns are referred to by an
expression! This moves those utilities into a new `analysis` module in
nom_sql, so they can be used in MIR as well.