Only send regression results to slack
We're running a lot of benchmarks now! Let's only send the results that
actually contain regressions to slack, skipping the ones that improved
or stayed the same.

We may want to include improvements here in the future, plus it might be
nice to have a web link of some sort to view more info