error handling: rewrite `sharding.rs` to fix `tpc_w` test
So the `tpc_w` test failed with an obscure sharding error, and redoing
the error handling refactor of `sharding.rs` makes it not do that.

I spent 2 hours on this and still have no idea why. This code is cursed,
and if you end up reading this commit message trying to figure out why
it broke again, I'm so sorry.