Make the debezium connected database type explicit.
This adds a new parmeter --database_type to the debezium connector
which may specify either "mysql" or "postgres". These two databases
vary in their messages for data collection naming and transaction
marker ids. By making this explicit we no longer have to infer
the database type from the message structure and the database type
should be known at deployment time.

In the future when the debezium connector may support multiple
databases we can refactor this code to infer the database types
more cleanly.