Add support for $1, $2 in collapse_where_in
This ticket adds support for postgres style placeholders in our
collapse_where_in logic. This is a necessary change for supporting a
postgres adaptor. The replacement works in accordance with the following
rules :

SELECT * FROM t WHERE id IN ($1, $2, $3) gets rewritten to
SELECT * FROM t WHERE id=$1

and
SELECT * FROM t WHERE x = $1 AND y IN ($2, $3, $4) OR z = $5 becomes
SELECT * FROM t WHERE x = $1 AND y=2 OR z = $5

Notice the final $5 is not turned into a lower number due to the
collapse.