Pull rhs columns through multiple filters
The function that builds the set of filter conditions for a MIR node
already had the ability to add projections for columns that were being
filtered on the *left hand side* of a condition expression out of the
parent node (by adding them to the columns in the filter node and then
letting pull_required_base_columns handle it later), but panicked if
it ever wasn't able to find a column on the *right hand side* of a
conditon expression. There was even a NOTE comment from jon expressing
suspicion about this! This commit pulls out the logic that adds the
column to a local helper function, and updates this to call it on both
sides of the condition.

No tests unfortunately right now, but this will be implicitly tested by
some integration tests to be added subsequently