Make referred_columns keep going until it's done
This fixes a bug in the iterator implementation for ReferredColumnsIter
where we could still have more things to traverse, but because we hit a
leaf on one side the overall iterator would think there was nothing left
to traverse. Now, instead, on each step of the iteration we keep looping
through all the various lists of things we want to traverse until we
either get another value or we know we have nothing left to traverse
because the lists are all empty.