Support explicitly parenthesized joins
SQL traditionally likes to think of the FROM clause of a query as a
full nested AST, where the various JOINs are binary operators with
tables or more JOINS on the left hand or right hand side. Certain ORMs -
notably SQLAlchemy in this case - emit SQL with *parentheses* around
JOIN clauses, to make the precedence of the operations explicit, and
previously we couldn't parse that - not least of which because our AST
just has a list of FROM tables and a list of JOIN clauses, with no
precedence present at all!

We *could* rework our AST to instead store the FROM clause as a tree,
but that would represent a pretty significant change to all the usage
sites, plus we're a little time-constrained atm, and about to do
a *different* AST refactor immediately after this block of changes.
Instead, this defines (a less-than-ideal version of) that tree AST as
an *intermediary* representation state for parsing the FROM clause, and
then unfold that AST into a list of tables and joins so we can then
store it in a SelectStatement. This is... not amazing, but works for now
until we can get a chance to do a larger refactor.