Run tests for noria server sub-crates in CI
It looks like we've been accidentally skipping running tests for the
sub-crates of noria/server in CI for quite some time - including some
rather important tests like reader_map and dataflow! This adds those as
targets to the "build noria" and "test noria" steps in the CI config,
and updates the tests to get them passing again.