Add support for compound joins in dataflow
Allow compound join keys in the Join ingredient, implemented by making
the `on` field on the Join type a *list* of pairs of indices rather than
just a single pair of indices. This ended up being a relatively easy
change - basically I just chased the type errors after that one change
and things worked right away.

This notably *doesn't* implement compound joins in MIR - there's still
an assertion that joins are a single comparison in QueryGraph that will
need to be removed to get this working fully through SQL.

Finally, I ran a benchmark of some joined queries to make sure the cost
of an extra pointer here both for storing the column indices and for
passing around the values for the join keys internally to the join
operator's on_input method itself wasn't too bad, and it definitely
isn't - if there's any change at all, it completely disappears into the
measurement noise.