Support keys with >6 columns
The `KeyedState` enum previously only let you have keys that were 6 or fewer
columns, mainly because having long tuples gets a bit unwieldy. We'd like to be
able to at least run queries with keys longer than that, even if they're slow.

This commit adds a `Multi` `KeyType` and `Multi{Hash, BTree}` `KeyedState` enum
variants in order to support this, which just use a `Vec<DataType>` as their
backing storage (with an additional field to denote the length).
We then remove the `mir_to_flow.rs` restriction of not having >6 group columns,
to unblock queries that require this.

Some extra stuff was added to `launchpad` to make working with `Bound`s easier,
as required for this commit. The current implementation clones a bit more than
I'd like it to, but so does everything else in the `KeyedState`, so...