Implement basic error handling for js-client
All errors encountered by the js-client _not_ caused by `Neon` are now
converted to instances of the javascript `error` class with the `message`
field set to the stringified version of the rust error.

We discussed serializing the rust errors into proper javascript error
objects with all of the fields propperly filled out,
and this is likely a more complete solution that provides the
js programmer with more error handling capability. However, currently
some of the errors handled by the client do not support easy
serialization, and even for those that do the serialization can fail and
then the failed serialization error case needs to be handled - producing
unpredictable error formats. Therefore, a more complicated solution has
been deferred until there is more motivation for it.