Implement compound joins in QG and MIR
Add full support for multi-column compound join keys all the way from
the query graph down through MIR, where it meets up with the
already-implemented compound join key support in dataflow.

A fair amount of the machinery was already here for this - most notably
the QueryGraphEdge::[Left]Join constructors already took a vector of
join predicates that were ANDed together, even though that vector was
only ever length 1. Weirdly, the JoinRef type that seems to have been
used for join order with regards to reuse had an `index` field for
specifically selecting one and only one of the join predicates out of
that query graph edge, but given that field would only ever select the
one and only predicate (and hence only ever was equal to 0) and there
isn't really much documentation around why it exists I can't figure out
what the intention was there - and all the tests still pass when it's
removed! So I removed it for now, and things still seem to work.

The rest of this commit is mostly mechanical - we run through all the
places that were assuming only one join key existed and translate them
to work with a list of join keys.