Refactor metric names to a mod of constants with documentation.
This change aggregates the metric names for all readyset counters
and gauges into the `recorded` module in `metrics.rs`. Each metric
should include: (1) what the metric is recording and (2) where the
metric is being recorded. This will allow us to easily parse the
set of metrics we are currently recording.