Timeout test steps after 60 minutes
Certain classes of test failure - notably those that cause the noria
server to crash but not the client - can cause the entire build process
to hang. I've seen cases where builds got stuck for multiple days,
locking up the build agent and clogging the queue so that other builds
couldn't make it through. This adds a timeout of 60 minutes (which we
basically always stay under) to the two main culprit steps - test noria
and test noria-mysql - so that these builds just fail instead.