Introduced a new MySQL TIME type implementation.
As part of the changes needed to get the MySQL built-in functions to work
(specifically, the `timediff` and `addtime` ones), we need to have a better
implementation of `Datatype::Time` (which must behave as a MySQL TIME type).
The problem with the current `Datatype::Time` implementation is that it relies
on a `NaiveTime` as its internal state; but a `NaiveTime` can only represent
timestamps within the range `00:00:00` to `23:59:59` (the possible times within
a day), whereas the MySQL TIME type can represent timestamps within the range
`-838:59:59` to `838:59:59`.
Currently, the `chrono` library only provides one utility to handle this kinds
of timestamps: the `Duration`. The problem is that `Duration` exposes very few
functions, which are not at all useful.
This commit introduces then the `MysqlTime` struct, which internally holds a
`Duration`, but wraps it in a way that let us control its behavior, to mimick
the MySQL TIME type.