Fix duplicated column name bug. ID 349.
Currently, when two tables have the same column name (such as id,) and a
comparison between the 2 is used to filter (such as t1.id >= t2.id,) the
filter node incorrectly creates the condition t2.id >= t2.id. This is
because the to_conditions function in MIR only compared names without
considering the table.

This is a very small fix and just adds a table comparison in the
to_conditions function. I also included an integration test with the
case described above.