Support function calls in filters
When generic function calls are done in a WHERE clause, current noria
fails to deal with this case (I think because the filter node's
evaluator can't handle it).

To get around this, we insert an extra project node before the filter,
using the new `project_expressions()` function. This projects out any
columns that contain function calls, and creates a lookup table that is
later used when constructing filter node conditions to use the new
projected columns.

We also extend the implied_tables SQL pass to rewrite function arguments
too (which is required for the dataflow lower-er to know what column is
being referred to in the function call).