Add columns to parents as needed by filters
When we're building up successive filter nodes from the set of
predicates for a query, sometimes those filters reference columns that
aren't necessarily projected out in the parents. Because unlike the rest
of MIR filter nodes use column indices instead of column names, this was
previously panicking when it tried to find the column in the parent. It
turns out that since we only really need to know the index in the
parent, it's okay to just add the column we need to the parent here and
then let pull_required_base_columns project it all the way through to
the parent afterwards. This is a bit of a hack, but will suffice until
we get a chance to refactor MIR filters to use column names instead of
indices (which we should do soon!).