Project with expressions shouldn't be query_through
There's an unstated invariant in noria that if an ingredient's
can_query_through method returns true, then all of that node's columns
must resolve to the parent. This invariant makes sense when you think
about how query_through works, which is that it takes a query for a
particular key to a node and sends that query along to the parent,
potentially applying some post-processing afterwards. If the query is
for a column that the node adds, then we obviously can't send that query
to the parent!

As best I can tell, the reason project used to be query_through is that
the first version of project didn't have the ability to evaluate
expressions or project literals - it could only ever reorder or omit
columns from its source node - and then expressions and literals were
added after the fact without considering the query_through invariants.
One thought I have is to disentagle this, and introduce a new "eval"
node that exists separately from project and is responsible for
evaluating expressions (of which literals are only one instance) and
which would get to be more intelligent about things like query_through.
In the meantime, though, this just only returns true for
can_query_through from project if the project node doesn't project out
any expressions.

I tried to reduce the query that triggered this bug to an
integration-test-sized reproduction case, but was unable to, so this is
untested right now - but when the weavegrid logictest eventually gets
the full query (should be soon) this will be tested by that.