Implement support for prepared statements in node.js client
Prepared statements are now prepared and executed via the
client.execute() method in the js client, which is modeled after the
mysql2 node.js library.

NOTE: mysql writes do NOT currently support execution with params. This
is being worked on now.

Other work to still be done on the client as of now:
- support conversion of timestamps between js & rust
- automated tests
- correct # of params checking for execution of prepared statements
- autoincrement sequence values
- documentation
- packaging and publication