Initial Packer pipeline
Initial Packer pipeline creation, at this time the only "flavor" we are
building is `zookeeper` which is the simplest.

An IAM policy with the ARN
`arn:aws:iam::069491470376:policy/BuildkitePackerBuilder` was created
for this and it was attached to the existing Buildkite role
`arn:aws:iam::069491470376:role/buildkite-Role`. The policy used the
minimum set of permissions as described in the
[docs](https://www.packer.io/docs/builders/amazon#iam-task-or-instance-role).

This PR includes the following:

- A Buildkite block step called `Build Packer images?` that allows a
 manual approval that triggers the Packer build process.
- A Packer lint step that makes sure all Packer files follow the
formatting standards (It runs `packer fmt`).
- A dynamic pipeline that checks for `.pkr.hcl` files inside
`ops/images` folder and generate the following steps for each file:
  - A `packer validate` step to make sure the Packer file is a valid
  one (This is ran in all branches, but `master` or `main`).
  - A `packer build` that will run the whole process but not
  publish the image, this is to validate we don't have a broken
  image procedure before merging to `master` or `main` branches.
  - Finally a "release" process that is only run in `master` or
  `main` branches that actually builds and publish the resulting
  images.

Failed build steps are retried 3 times (This is because sometimes
Packer fails either due to network or timeout issues, so it is better
if it is automatically retried).

The resulting AMI will have the following name syntax:

`zookeeper-YYYYMMDD-(short_sha_commit_id)`

Where `YYYYMMDD` is the current date expressed in that format
and `short_sha_commit_id` is taken from `SHORT_COMMIT_ID`. If
Packer is run outside the Buildkite pipeline it will default
to `0000000`.

Additionally, the resulting AMI will be available in `us-east-1` and
`us-east-2` regions (More can be added in the Packer file if needed).

Once this PR is merged other "flavors" will be added to the pipeline.