Changed inner representation of Datatype::Time
Previously, `Datatype::Time` used a `NaiveTime` as its internal representation.
However, a `NaiveTime` is unable to represent a MySQL TIME type correctly, so
this commit changes the `Datatype::Time` to use `MysqlTime` as its internal
representation instead.
Note that it was necessary to have custom implementations of `FromValue`,
`From<MysqlTime> for myc::value::Value`, `ConvIr<MysqlTime> for
ParseIr<MysqlTime>` and `ParseIr<T>`, since these were implemented for
`NaiveTime` in the `mysql_common` crate but now where needed for `MysqlTime`.
Sadly, `mysql_common` crate does not expose the tools to implement these in a
clean one. For that reason, the struct `ParseIr` was copy-pasted from the
library in order to implement the other traits.