Add columns to filters in the right order
Filter nodes (perhaps uniquely?) aren't capable of reordering the set of
input columns - this means that when adding columns as part of
pull_required_base_columns, if a filter node is an immediate child of a
node that has *rules* about where it adds columns - for example an
aggregate or projected column - the column will get added to the filter
in *a different place* than it gets added to the parent node, causing
the column to be rearranged.

This updates the add_column method on MirNode to add columns to filters
in the same place as those columns are added in the parents, potentially
recursing up multiple parents until we hit a node that isn't a filter.
This isn't *sufficient* to make this work properly, since we
might also add a column in a place that causes a filter *condition* to
get reordered, resulting in that condition pointing at the wrong index -
fixing that is next.