Update Rust toolchain version
The previous toolchain version wasn't being served by rustup any more
(I found this out when it failed to download on a new machine). This new
version seems to be both available in rustup and compatible with the
codebase as-is; I encountered problems compiling everything on latest
nightly, which we'll have to fix soon.

We re-run rustfmt with the new toolchain version, and get rid of
warnings for things that got stabilized.