Add an external request to clear metrics at the noria-server.
This adds support for a new http endpoint /metrics_reset that
allows an external client to reset the state of the
NoriaMetricsRecorder. This functionality will allow our benchmarking
service to reset the metrics controller between runs.

Long-term this is not a good solution as currently anyone can reset
metrics at one of our servers through an HTTP request. It will need
to be changed when we introduce more security mechanisms.