error handling: refactor noria_connector.rs partially
I missed out the `NoriaConnector` in the initial error handling PR, and
it has a bunch of asserts & such that users are likely to hit (in fact,
I only did this because we did hit one while I was pairing with grfn!).

This is a fairly rudimentary s/panic/internal/-type commit, with the
exception of the `ValueCoerceError` hack (which is a separate error type
that should be included in the main `ReadySetError` eventually, but I
didn't want to rip it out because that'd be a larger change).