Reorder filter conditions on add_column
Adding columns to filter nodes might happen at something other than the
last column index, depending on the rules for adding columns to the
parent nodes of the filter - see 561134fdc (Add columns to filters in
the right order). Since filters are (still) index-based rather than
name-based, it might be the case that after following these rules the
column indices referenced in filters might need to be updated, if
they're pushed forward as the result of adding a new column. This makes
that happen, along with some unit tests to check that it behaves
correctly.

I've said this before (likely recently, in the log), but I *really* want
to refactor MIR filters to be name-based rather than index based so this
doesn't have to happen.