Only dedupe join keys for inner joins
When building join nodes, MIR previously deduplicated the output column
for join keys, since they exist on both the left-hand and right-hand
side of the join. This is only correct behavior on *inner* joins, since
the query might still reference the now-nullable right-hand side of the
join key after the join for a left join, for instance as part of a
filter to build an outer join. This commit changes that to only do the
join key deduplication + aliasing on inner joins, and covers it with an
integration test that performs an outer join.