Follow the column rules on the rhs of a filter
Commit 561134fdc (Add columns to filters in the right order), which
corrected the implementation of MirNode::add_column for filter nodes to
respect the column ordering rules of the filters' parent node, only
applied to the places that actually *called* add_column for filters -
notably the left-hand side of to_conditions and the
pull_required_base_columns post-mir pass. The *right hand* side of
filters had its own, incomplete pattern match on the node type of the
parent to partially reimplement the column ordering rules, which didn't
properly handle project nodes (or filteraggregates, or another filter in
front of another node with a column ordering rule). This meant that
columns added on the right hand side of a filter wouldn't properly
respect the column ordering rules of the parent and would end up mixing
up columns. This changes the right-hand-side column adding logic to
instead just call add_column on the parent and use the index returned
there to add the column to the filter, which while still a bit hacky
(we should really just refactor this to be name-based, again) gets us to
a place where filters shouldn't be getting columns mixed up anymore