Pass NULLs through ProjectExpression::eval
The semantics of SQL state that NULL is infectious - if a function like
eg MONTH is called on a NULL value the expected result is NULL.
Previously, we were calling `into()` on the `DataType` returned by
`coerce_to` in the eval implementation for certain functions, which was
resulting in a `panic` if that DataType was null. This adds a new macro
`non_null`, which instead of panicking checks if its argument is null
and short-circuit returns DataType::None if so.