SQL CHAR types accept too-short values
My previous implementation of coerce_to for the CHAR type was too
restrictive: it failed on inputs that were shorter than the given
length. Per the documentation[1], what actually happens in MySQL proper
is CHAR fields are padded up to the given length, then stripped of
spaces on retrieval (unless PAD_CHAR_TO_FULL_LENGTH is set). This is a
little complex for us to introduce right now, so I've elected to instead
just treat the length for CHAR as a maximum length and skip the padding
altogether, which assuming the above setting is off and given we're not
actually using the passed length to allocate space for the field has the
same user-facing effect.

[1]: https://dev.mysql.com/doc/refman/8.0/en/char.html