Allow sharding by timestamp fields
If shard_by in noria is passed a Timestamp Datatype, rather than
panicking, just convert the timestamp into a text and hash that to use
as the shard key. This gets us a good source of entropy for the shard
key, since due to natural clustering a numeric field might end up being
unevenly balanced. At some point in the future we may want to benchmark
this specific implementation and replace it with something that hashes
the individual fields of the timestamp rather than allocating a string
first, but for now this should be fine.