Attempt to infer the types of project expressions
The noria-client needs to know the types of the result columns of a view
so it can communicate those to the client as part of the prepare query
result, and use them to serialize the results of the query itself.
Previously, the `schema` for a view always assumed the result of a
project expression was a BIGINT, regardless of what the expression
actually did, which meant that project expressions returning other
types (CAST AS DATETIME), CONVERT_TZ, etc. etc. would panic when
serializing results in noria_client. This commit adds some *extremely*
rudimentary type inference to the ProjectExpression type, essentially
just hooking into the bit in schema that's tracing columns back into
their origin for the expressions like ifnull whose type depends on their
input type.

There are a *lot* of limitations to the current approach - for one, we
basically assume everything is actually well-typed (this is just a type
inference engine, not a typechecker) - for another, we don't do *any*
unification even where we should (eg the arguments of IFNULL or the
operands of a binary operator that could be either on floats or ints)
which requires tons of explicit CASTS (essentially type ascription) in
queries to get the results of columns to actually infer to the right
type. Very soon, we should rip out this current implementation entirely
and replace it with a holistic whole-query typechecker which attaches
type ascription to the entire query graph at the MIR level.