Don't shuffle grouped columns in aggregate nodes
For some reason - and we can't really figure out why - the grouped
ingredient was *sorting* the grouped column indices from the parent, and
then building a "colfix" vector to look up into in the implementations
of resolve and parent_columns. This resulted in correct replay paths and
column resolution, but also (unsurprisingly?) resulted in the grouped
nodes themselves completely mixing up the columns in their output, since
the reordering of the grouped columns never got communicated to the
actual query graph itself. Again, without understanding *why* the
grouped columns were sorted in the first place (the original commit
message on the line doesn't really tell us much) removing the sort
initially doesn't appear to *hurt* anything, and fixes the issue where
the output columns are getting mixed up.