Add date and time support to logictest
Add types and values to logictest corresponding to DataType::Date and
DataType::Time. Since "T" is already taken by Text, the letter for the
Time type is "M" - I considered just extending it to multiple letters,
but decided that the visual simplicity of the single-letter thing was
better than the slightly more obvious type.