Add a basic script generator to logictest
Add the beginning of support to logictest for *generating* test scripts,
by running a series of queries against mysql and saving those results to
a new script.

A new `generate` subcommand takes a "seed" script containing CREATE
TABLE statements and queries, and parses those CREATE TABLE statements,
using them to initialize a query_generator::GeneratorState, which is
then used to generate a series of INSERT statements for data into those
tables. We then run the CREATE TABLE statements and INSERT statements
against a passed-in reference mysql database, and use the results to
generate new QUERY records in a new test script that is generated as
output. To get both primary keys and joins working, there's a new
`unique_columns` field in GeneratorState, which specifies columns that
should *always* receive unique values.

In the future, I'd like to make the generation of insert statements here
optional, so that when we *do* actually have our own data we can just
run queries against mysql and save the results to a test script.