Always run Materializations::setup() for new indices
...(even for nodes that were unmaterialized and have now become fully
materialized)

This seems so obvious that I feel like I'm probably missing something.

So, the context is as follows: the materialization planner gets run when
new indices are added to nodes, in order to figure out what replay paths
need to be either created or used for the new indices. In the case of
full materialization, the planner will schedule a chunked replay through
these paths when it's finalized (and Materializations::setup goes and
does the work of actually executing said replay) -- which is, of course,
necessary to make sure that they actually are fully materialized.

However, the current code only did this when *new* fully materialized
nodes were created. It's possible for the code to turn nodes from
completely not materialized into fully materialized. In this case, the
materialization planner is *NOT* run, and the existing node that just
got made fully materialized is initialized with empty indices -- which
is a bug, because the node never has data replay into it!

This caused, for example, a real-world weavegrid bug wherein no records
were returned from a query, because they were never replayed into the
necessary dataflow nodes because the planner never ran.

To remedy this, we now *always* run the planner, for *every* new added
index. In [fully materialized] -> [fully materialized] cases, this will
just have the effect of adding new indices for the existing data, as
before, and is arguably somewhat more computationally wasteful. This
makes the code somewhat easier to reason about, though, as it is no
longer necessary to reason about the case where the planner is not run
-- everything replay-related happens in plan.rs, and everything
index-related happens in mod.rs.

(This requires tweaking the planner to reject replay paths that are <=1
node in length, otherwise some parts of the code get confused.)