Add CI steps to build release binaries
Add steps at the end of the buildkite pipeline, after a block step but
before the packer image build step, to build release binaries for
noria-server, noria-mysql, and debezium_connector, and upload them as
artifacts. Also, since the rest of the steps that *used* to be building
binaries are no longer important for anything except for
typechecking (we don't do anything with the binaries that get built
there), this converts those steps to `cargo check` instead of `cargo
build`, which should speed things up in the case of a non-release build.