Move parse error -> 0 coercion out of MysqlTime
Certain things (noria-logictest, here) need to know the difference
between a mysqltime that failed to parse and one that is actually just
all 0s - specifically for noria-logictest we need proper parse failures
so we can try the next branch in the `alt!` we use to parse expected
return values. We still want to replicate MySQL's behavior where failed
parses result in an all-0 time, though (and we need to be able to tell
the difference between a failure for that reason and a failure because
one of the fields was out-of-bounds), so this defines a new error enum
for failing to convert strings to MysqlTimes so we can differentiate,
then does that differentiation in DataType::coerce_to when parsing
strings as times.