Make the materialization planner (more) idempotent
Given we run the materialization planner far more frequently now, it's a
good idea to make sure it's idempotent (that is, rerunning it isn't
going to result in extra replay paths or indeed replays being added /
done!).

This commit makes `Materializations` keep track of added replay paths
every time the planner is run, and makes the planner use this
information to avoid adding indices to fully materialized nodes that had
indices added by a previous run of the planner. That should ensure that
no double-replaying happens in the case of full materialization (which
was a risk with the 'always run the planner' commit).

Making sure double-replaying is impossible under partial materialization
is a separate challenge that is left to a later date (given that the
'always run the planner' commit didn't change the behaviour under
partial, given it always *was* run under partial).

I don't think double-replaying occurs currently under partial, but we
want to check for that in future.