Wait for http requests, not just TCP connections
The debezium-connect docker image starts accepting tcp connections
before it starts actually accepting http requests to those tcp
connections - this meant that waiting for /dev/tcp/localhost/8083 didn't
mean that a subsequent curl request to create the connector would
succeed. This changes the wait script to instead wait for an HTTP
request to localhost:8083 to succeed, which fixes the issue where the
container isn't getting created.