Teach MemoryState to check overlapping indices on miss
97fc5eb2c3272f3d8a82f94d24df394bbd70e3e5 introduced the "partially
overlapping partial indices" check, and provided a variety of potential
solutions in its commit message. This commit implements the "more robust
solution", which involves teaching Noria about overlapping indices.

[ begin problem recap ]

When we have a parent node indexed on columns [0, 1], but a child node
indexed on [1], it was previously possible to get into "really sad
situations" where the child might upquery along the [1] replay path.
This seems fine, except when the parent processes the records returned
via replay, it'll end up doing lookups in its [0, 1] index for them --
which will miss, since it hasn't been marked filled by the replay logic.
This would necessitate unnecessary extra upqueries -- which is just a
performance problem -- but would *also* mean that future queries to the
[0, 1] index would incorrectly miss.

This is because the parent node would only upquery for the values of
column 0 relevant to the records being processed; when a record flowed
through the parent during normal processing that had a different value
of 0, but the same value of 1, the parent would incorrectly drop the
record due to a miss, and not propagate the state to the child node
(which should receive it, given it has all the values for column 1!),
resulting in the child having stale state. This is bad!

[ end problem recap ]

To work around this, then, when we miss on an index lookup, we also
consult all other indices whose columns are a strict subset of the
missed index's columns (e.g. [0] and [1] for [0, 1]). Some probably
quadratic logic remaps the columns of the lookup key to fit the new
index, which is then queried; if it returns results, we filter the
results to make sure they actually fit the key and return them.

We nerf the "partially overlapping partial indices" check to only fire
if the above logic would not trigger -- which I think is "never", given
that I'm pretty sure you're not allowed to have an index in a child that
isn't also in a parent. That check can and probably should be removed at
a later date.

KeyType::get() and RecordResult::retain() methods were added to support
the above usecase.

The new `overlapping_indices` integration test exercises the new logic.
Without the new logic enabled, it would return an empty set of records;
with the new logic, it returns a correct result!